added more ui tests
